retriever/requirements.txt
fastapi==0.100.0
loguru==0.7.0
uvicorn==0.29.0
python-dotenv==1.0.1
pinecone==5.4.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.31.0
python-multipart==0.0.9
pillow==10.4.0
# gradio==4.44.1
gradio==4.19.2
